requirements.txt
diff-match-patch==20181111
Django==2.1.3
django-filter==2.0.0
django-import-export==1.1.0
django-tinymce4-lite==1.7.3
django-uuslug==1.1.8
django-webpack-loader==0.6.0
djangorestframework==3.9.0
et-xmlfile==1.0.1
jdcal==1.4
jsmin==2.2.2
Markdown==3.0.1
odfpy==1.3.6
openpyxl==2.5.11
Pillow==5.3.0
python-decouple==3.1
python-slugify==1.2.6
pytz==2018.7
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
Unidecode==1.0.23
xlrd==1.1.0
xlwt==1.3.0
